Extract search field
