+++ lib/widgets/inputs/multiselect_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:spare_parts/pages/home_page/inventory_view/multiselect_dialog.dart';
+import 'package:spare_parts/widgets/inputs/multiselect_dialog.dart';
 
 class MultiselectButton extends StatelessWidget {
   final List<String> values;
